Make all optional fields computed=true
The basic idea is the following:

- if the property is not specified in the tf file, terraform should just
  leave it alone and not try to modify it.

Example:

Imagine a resource is defined in tf and the customer decided not
to configure `require_mfa_to_approve` in tf. If `require_mfa_to_approve`
was optional and not computed, terraform apply would try to change
it back to the value it had when it was created (i.e. false) if someone
changed the setting in the UI.

In a world where these attributes are optional + computed, we're
basically saying, terraform should leave anything that the user does
not explicitly specify alone. That seems to be easier to reason about
for the time being.