{group,resource}: fix bug where auto approve on create did not work
After we changed this to always require admin owner id, the else if
code block would always run. This was an undetected bug that lurked for
a while until we changed the API to be better about keeping auto approval
settings in sync with reviewers, which uncovered the buggy behavior.